Additional refactoring around the time sheet summary
This commit refactors the logs for summarizing the time sheet into a function called `summarizeTimeSheet`, which can be used where ever we need. I also converted a few of our `setIntervalAsync` calls to `while` loops, since this reduces nesting.
